% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_bibtex.R
\name{export_bibtex}
\alias{export_bibtex}
\title{Create BibTeX citations for the requested datasets.}
\usage{
export_bibtex(uuids, outfile, append = FALSE)
}
\arguments{
\item{uuids}{Character vector of UUIDs corresponding to
the desired datasets.}

\item{outfile}{Character vector of length 1. The path to
the file the citations should be written to.}

\item{append}{Logical vector of length 1. If TRUE, the
outfile will be appended to. If FALSE, the outfile will
be overwritten.}
}
\description{
Resolves the DOI for each requested dataset and exports
the resulting list of BibTeX citations to outfile. Does
not export duplicate citations.
}
