% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_batch_key.R
\name{get_batch_key}
\alias{get_batch_key}
\title{Get the "batch_key" attribute.}
\usage{
get_batch_key(uuid)
}
\arguments{
\item{uuid}{Character vector of length 1. UUID of the desired dataset.}
}
\value{
A character vector of length 1 containing the batch_key
}
\description{
Get the "batch_key" attribute from the cell-specific
internal universal metadata field "batch" for the
indicated dataset.
}
