% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bibtex.R
\name{get_bibtex}
\alias{get_bibtex}
\title{Get the BibTeX citation for an entry.}
\usage{
get_bibtex(uuid)
}
\arguments{
\item{uuid}{Character vector of length 1. The UUID of the
desired dataset.}
}
\value{
A character vector of length 1. It contains the BibTeX
  entry as a printable string.
}
\description{
Resolves the DOI for a dataset to generate a BibTeX citation
as a string.
}
