% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_author_annot.R
\name{get_cell_author_annot}
\alias{get_cell_author_annot}
\title{Get the cell author-annotations.}
\usage{
get_cell_author_annot(uuid)
}
\arguments{
\item{uuid}{Character vector of length 1. UUID of the desired dataset.}
}
\value{
A dataframe holding all the data available from the cell
  author-annotations.
}
\description{
Get the cell author-annotations from a given dataset
as a dataframe.
}
