% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_ids.R
\name{get_cell_ids}
\alias{get_cell_ids}
\title{Get the cell IDs for a given dataset.}
\usage{
get_cell_ids(uuid)
}
\arguments{
\item{uuid}{Character vector of length 1. UUID of the desired dataset.}
}
\value{
A character vector containing the cell IDs.
}
\description{
For a given dataset, retrieve the cell IDs.
}
