% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cell_univ.R
\name{get_cell_univ}
\alias{get_cell_univ}
\title{Get the cell universal metadata.}
\usage{
get_cell_univ(uuid, keep_missing = TRUE)
}
\arguments{
\item{uuid}{Character vector of length 1. UUID of the desired dataset.}

\item{keep_missing}{Logical vector of length 1. If TRUE, the returned
dataframe will have all columns. If FALSE, columns where all values
== -1 or "-1" (indicating a missing column) will be dropped from
the returned dataframe.}
}
\value{
A dataframe holding all the data available from the cell
  universal metadata.
}
\description{
Get the cell universal metadata from a given dataset
as a dataframe.
}
