% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column_allmissing.R
\name{get_column_allmissing}
\alias{get_column_allmissing}
\title{Get the all_missing attribute for a given column.}
\usage{
get_column_allmissing(uuid, column, var = "cell", metadata = "universal")
}
\arguments{
\item{uuid}{Character vector of length 1. UUID of the desired dataset.}

\item{column}{Character vector of length 1. The name of the column, for
which the "all_missing" attribute should be retrieved.}

\item{var}{Character vector of length 1. Must be "cell" or "gene".
Indicates which variable's metadata the column is in.}

\item{metadata}{Character vector of length 1. Must be "universal" or
"author_annot". Indicates which metadata the column is in.}
}
\value{
A character vector of length 1, holding the HDF5 "all_missing"
  attribute of the indicated column. If there is no all_missing
  attribute available, a message is printed, and the function returns NULL.
}
\description{
For a given internal metadata column, retrieve the HDF5
"all_missing" attribute, if available, from the requested
dataset.
}
