% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expr_mat.R
\name{get_expr_mat}
\alias{get_expr_mat}
\title{Get an expression matrix from a dataset.}
\usage{
get_expr_mat(uuid, matrix = "matrix")
}
\arguments{
\item{uuid}{Character vector of length 1. UUID of the desired dataset.}

\item{matrix}{Character vector of length 1. Name of the desired expression
matrix. If "matrix", the main expression matrix under lfile[["matrix"]]
will be returned. Otherwise, must be a valid layer.}
}
\value{
A genes x cells unnamed sparse matrix holding the requested
  expression matrix. It is a Matrix::dgCMatrix class.
}
\description{
Get the named expression matrix from a dataset. Default
is to get the expression matrix stored at lfile[["matrix"]],
but any of the named layers can be gotten.
}
