% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_expr_mat_names.R
\name{get_expr_mat_names}
\alias{get_expr_mat_names}
\title{Get the expression matrix names.}
\usage{
get_expr_mat_names(uuid)
}
\arguments{
\item{uuid}{Character vector of length 1. UUID of the desired dataset.}
}
\value{
A character vector containing the names of the matrices.
}
\description{
Get the expression matrix names from a dataset. The first
entry will always be "matrix", referring to the expression
matrix stored under lfile[["matrix"]], not under
lfile[["layers"]].
}
