% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extern_md.R
\name{get_extern_md}
\alias{get_extern_md}
\title{Get the external metadata.}
\usage{
get_extern_md()
}
\value{
A dataframe containing the external metadata.
}
\description{
Get the external metadata as a dataframe.
}
