% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_author_annot.R
\name{get_gene_author_annot}
\alias{get_gene_author_annot}
\title{Get the gene author-annotations.}
\usage{
get_gene_author_annot(uuid)
}
\arguments{
\item{uuid}{Character vector of length 1. UUID of the desired dataset.}
}
\value{
A dataframe holding all the data available from the gene
  author-annotations.
}
\description{
Get the gene author-annotations from a given dataset
as a dataframe. Sets rownames preferentially as Accession,
Gene, or a number range in that order. Accession and Gene
will not be used if they have non-unique values.
}
