% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_ids.R
\name{get_gene_ids}
\alias{get_gene_ids}
\title{Get the gene IDs for a given dataset.}
\usage{
get_gene_ids(uuid, accession = TRUE)
}
\arguments{
\item{uuid}{Character vector of length 1. UUID of the desired dataset.}

\item{accession}{Logical vector of length 1. If TRUE, the "Accession"
variable will be returned. If FALSE, the "Gene" variable will be
returned. If the requested variable is not available, but the other
is, the other will be returned with a warning.}
}
\value{
A character vector containing the gene IDs.
}
\description{
For a given dataset, retrieve the gene IDs.
}
