% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_univ.R
\name{get_gene_univ}
\alias{get_gene_univ}
\title{Get the gene universal metadata.}
\usage{
get_gene_univ(uuid, keep_missing = TRUE)
}
\arguments{
\item{uuid}{Character vector of length 1. UUID of the desired dataset.}

\item{keep_missing}{Logical vector of length 1. If TRUE, the returned
dataframe will have all columns. If FALSE, columns where all values
== -1 or "-1" (indicating a missing column) will be dropped from
the returned dataframe.}
}
\value{
A dataframe holding all the data available from the gene
  universal metadata.
}
\description{
Get the gene universal metadata from a given dataset
as a dataframe. SEts rownames preferentially as Accession,
Gene, or a number range in that order. Accession and Gene
will not be used if they have non-unique values.
}
