% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_h5_conn.R
\name{get_h5_conn}
\alias{get_h5_conn}
\title{Get an hdf5r connection to a loom file.}
\usage{
get_h5_conn(uuid, warning = TRUE)
}
\arguments{
\item{uuid}{Character vector of length 1. UUID of the desired dataset.}

\item{warning}{Logical vector of length 1. If \code{TRUE}, a long
warning will be displayed, reminding the user that R accesses
the loom file in column-major order, but that the file is stored
in row-major order. Thus all accessed matrices will be transposed.}
}
\value{
H5File connection to the desired loom file.
}
\description{
\code{get_h5_conn()} returns an hdf5r H5File
connection object to the loom file associated
with the passed UUID.
}
\details{
The only mode is read-only.
}
