% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loom_conn.R
\name{get_loom_conn}
\alias{get_loom_conn}
\title{Get a loomR connection to a loom file.}
\usage{
get_loom_conn(uuid, warning = TRUE)
}
\arguments{
\item{uuid}{Character vector of length 1. UUID of the desired dataset.}

\item{warning}{Logical vector of length 1. If \code{TRUE}, a long
warning will be displayed, reminding the user that R accesses
the loom file in column-major order, but that the file is stored
in row-major order. Thus all accessed matrices will be transposed.}
}
\value{
loomR connection to the desired loom file.
}
\description{
\code{get_loom_conn()} returns an loomR
connection object to the loom file associated
with the passed UUID.
}
\details{
The only mode is read-only.
}
