% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loom_filename.R
\name{get_loom_filename}
\alias{get_loom_filename}
\title{Get a loom filename.}
\usage{
get_loom_filename(uuid)
}
\arguments{
\item{uuid}{Character vector of length 1. UUID of the desired dataset.}
}
\value{
Character vector of length 1 containing the full path to
  the desired loom file.
}
\description{
\code{get_loom_filename()} returns the full filepath to the
loom file associated with the passed UUID.
}
\details{
This filepath is retrieved from the external metadata file.
If the file does not exist, an error will be thrown.
}
