% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_note.R
\name{get_note}
\alias{get_note}
\title{Get the note for a dataset.}
\usage{
get_note(uuid)
}
\arguments{
\item{uuid}{Character vector of length 1. The UUID of the
desired dataset.}
}
\value{
A character vector of length 1. If the UUID has
  a note, the note is returned. If not, an empty string
  is returned.
}
\description{
Retrieves the note for a dataset from the "notes.tsv"
file at the root of the database. These are typically
small notes used when an entry is still in progress or
being edited in the future.
}
