% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PATH_TO_DATABASE.R
\name{get_PATH_TO_DATABASE}
\alias{get_PATH_TO_DATABASE}
\title{Get the path to the root of the database}
\usage{
get_PATH_TO_DATABASE()
}
\value{
A character vector of length 1 holding the full
  path to the root directory of the database.
}
\description{
Gets the full file path to the root of the database.
Takes into account the differential naming on dactyl.
}
