% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_PATH_TO_METADATA.R
\name{get_PATH_TO_METADATA}
\alias{get_PATH_TO_METADATA}
\title{Gets the path to the external metadata.}
\usage{
get_PATH_TO_METADATA()
}
\value{
A character vector of length 1 holding the full
  path to the external metadata file.
}
\description{
Gets the full filepath to the external metadata file.
Takes into account the different naming on dactyl.
}
