% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sce.R
\name{get_sce}
\alias{get_sce}
\title{Get a SingleCellExperiment object from a loom file.}
\usage{
get_sce(
  uuid,
  assay_for_matrix = "counts",
  counts_assay = NULL,
  logcounts_assay = NULL
)
}
\arguments{
\item{uuid}{Character vector of length 1. UUID of the desired dataset.}

\item{assay_for_matrix}{Character vector of length 1. The name of the
assay that the 'matrix' dataset should be stored in.}

\item{counts_assay}{Character vector of length 1, or NULL. If not NULL,
must be the name of an HDF5 dataset under 'layers' that will be
stored under the "counts" assay of the returned SingleCellExperiment
object. For obvious reasons, this argument cannot be not NULL if
\code{assay_for_matrix} == "counts".}

\item{logcounts_assay}{Character vector of length 1, or NULL. If not NULL,
must be the name of an HDF5 dataset under 'layers' that will be
stored under the "logcounts" assay of the returned SingleCellExperiment
object. For obvious reasons, this argument cannot be not NULL if
\code{assay_for_matrix} == "logcounts".}
}
\value{
SingleCellExperiment object with the assays as assigned and the
  'col_attrs' datasets consolidated into
  \code{SingleCellExperiment::colData()} and the 'row_attrs' dataset
  consolidated into \code{SingleCellExperiment::rowData()}.
  Additionally, the batch_key, and author annotated metadata are
  stored in the \code{S4Vectors::metadata()} slot, named "batch_key",
  "cell_author_annot", and "gene_author_annot".
}
\description{
\code{get_sce()} returns a SingleCellExperiment object
containing the entire loom file loaded completely into
memory.
}
\details{
The loom file associated with the UUID is the one
loaded into an SCE object. The other 3 function arguments
allows any configuration of SingleCellExperiment assays.
}
\section{Assay assignment}{

The loom file has a main matrix in the HDF5 dataset 'matrix'.
The rest of the matrices, if available, are under the
HDF5 group 'layers', as datasets of their own. SingleCellExperiment
objects typically have the raw counts stored under the assay "counts",
and the log-normalized counts stored under the assay "logcounts".
Other matrices can be named anything. The \code{assay_for_matrix}
argument should hold the name of the assay that the 'matrix' dataset
should be stored in. Then, the \code{counts_assay} and
\code{logcounts_assay} arguments can hold the names of HDF5 layers
that should be stored in "counts" or "logcounts" assays respectively.
For example: if the loom file has a raw reads matrix in 'matrix'
and a normalized matrix in 'norm' and a 3rd matrix in 'misc', the
appropriate function call would be:
\code{get_sce(UUID,
              assay_for_matrix = "counts",
              counts_assay = NULL,
              logcounts_assay = "norm")}
However, if the 3rd matrix was in 'matrix' and the raw reads and
normalized reads were in 'raw' and 'norm', the appropriate function
call would be:
\code{get_sce(UUID,
              assay_for_matrix = "misc",
              counts_assay = "raw",
              logcounts_assay = "norm")}
}

