% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sce.R
\name{merge_sce}
\alias{merge_sce}
\title{Merge SingleCellExperiment objects returned from the database.}
\usage{
merge_sce(
  sce1,
  sce2,
  keep_author_annot = FALSE,
  min_common_genes = 15000,
  cell_id_prefix = 4
)
}
\arguments{
\item{sce1}{SingleCellExperiment object. The first
SingleCellExperiment object to be merged. Should be an
object returned from \code{get_sce()}.}

\item{sce2}{SingleCellExperiment object. The second
SingleCellExperiment object to be merged. Should be an
object returned from \code{get_sce()}.}

\item{keep_author_annot}{Logical vector of length 1. If TRUE,
the internal author-annotated metadata from each SCE will
be labeled with cell_id_prefix and saved in the metadata
of the new SCE. Otherwise, they will be dropped.}

\item{min_common_genes}{Numeric vector of length 1. Must
be an non-negative integer. The two SCE objects must
have this many valid gene IDs in common. Otherwise, the
merge will fail.}

\item{cell_id_prefix}{Can be a single integer from the
following list \code{c(4, 8, 12, 16, 20, 24, 28, 32)} or
a character vector of length 2 with 2 different prefixes
for the cell labels from \code{sce1} and \code{sce2}
respectively. If it is an integer, that number of digits
from the UUIDs are used as a prefix (does not include the
hyphens). If the number is too small to produce unique
prefixes, it is increased by 4 until it works.}
}
\value{
A new SingleCellExperiment object containing all
  cells from both objects (\code{sce1} first). It only
  retains genes that were guaranteed to be unique inside
  each SCE object and in common between the two objects.
  Cell IDs are prefixed with \code{cell_id_prefix}. The
  \code{colData()} from each is also merged. The names of
  the author-annotated internal metadata dataframe, if
  kept are prefixed with the \code{cell_id_prefix}. The
  function attempts to combine the batch_keys into a 
  vector, but may not do this correctly if sce1 or sce2
  are the product of a previous merge.
}
\description{
Attempts to merge 2 SingleCellExperiment objects returned from
get_sce(). Will reject the merge if any incompatibilities
exist.
}
