% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid_to_row.R
\name{uuid_to_row}
\alias{uuid_to_row}
\title{Get an entry of the external metadata.}
\usage{
uuid_to_row(uuid, columns = NULL)
}
\arguments{
\item{uuid}{Character vector of length 1. UUID of the desired dataset.}

\item{columns}{Character vector containing requested columns from
the external metadata. If NULL, the entire entry will be returned.
If a column is invalid, the function will still run, however the
entry will contain "NOT A VALID COLUMN" in lieu of the actual value.
Additionally, a warning will be thrown.}
}
\value{
A dataframe holding one row (the entry) where the column
  names match the columns argument.
}
\description{
Get the entry of the external metadata corresponding
to the requested UUID.
}
